feat: add snowflake layer, refactor cdk define
