+++ lambdas/load_config.py
@@ -1,5 +1,4 @@
 from typing import Dict
-# import snowflake.connector
 
 
 def handler(event: Dict, context):
