+++ lambdas/requirements.test.txt
@@ -1,6 +1,30 @@
+asn1crypto==1.5.1
+boto3==1.35.88
+botocore==1.35.88
+certifi==2024.12.14
+cffi==1.17.1
+charset-normalizer==3.4.1
 coverage==7.6.10
+cryptography==44.0.0
+filelock==3.16.1
+idna==3.10
 iniconfig==2.0.0
+jmespath==1.0.1
 packaging==24.2
+platformdirs==4.3.6
 pluggy==1.5.0
+pycparser==2.22
+PyJWT==2.10.1
+pyOpenSSL==24.3.0
 pytest==8.3.4
 pytest-cov==6.0.0
+python-dateutil==2.9.0.post0
+pytz==2024.2
+requests==2.32.3
+s3transfer==0.10.4
+six==1.17.0
+snowflake-connector-python==3.12.4
+sortedcontainers==2.4.0
+tomlkit==0.13.2
+typing_extensions==4.12.2
+urllib3==2.3.0
