+++ layers/requirements.txt
@@ -1,0 +1,18 @@
+asn1crypto==1.5.1
+certifi==2024.12.14
+cffi==1.17.1
+charset-normalizer==3.4.1
+cryptography==44.0.0
+filelock==3.16.1
+idna==3.10
+platformdirs==4.3.6
+pycparser==2.22
+PyJWT==2.10.1
+pyOpenSSL==24.3.0
+pytz==2024.2
+requests==2.32.3
+snowflake-connector-python==3.12.4
+sortedcontainers==2.4.0
+tomlkit==0.13.2
+typing_extensions==4.12.2
+urllib3==2.3.0
